+++ Kilometros Database/UserDataTotalStepsView.cs
@@ -17,6 +17,6 @@ namespace KilometrosDatabase
         public System.Guid User_Guid { get; set; }
         public short Activity { get; set; }
         public Nullable<System.DateTime> Timestamp { get; set; }
-        public Nullable<int> TotalSteps { get; set; }
+        public Nullable<long> TotalSteps { get; set; }
     }
 }
